LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/top-layer-nested-dialog-expected.txt


PASS Dialogs initially not matching for container queries
PASS Dialogs still not matching after showModal
FAIL @container queries start matching assert_equals: expected "rgb(0, 255, 0)" but got "rgb(255, 0, 0)"
